+++ server-utils/src/access_control.rs
@@ -1,13 +1,64 @@
-use crate::hosts::{AllowHosts,Host};
+//! Access control based on http headers
+use crate::{cors,hosts};
 use crate::cors::{AccessControlAllowHeaders,AccessControlAllowOrigin};
+use crate::hosts::{AllowHosts,Host};
+use hyper::{self, header};
+use crate::utils;
 
+
+/// Define access on control on http layer
 #[derive(Clone)]
 pub struct AccessControl {
-    pub allow_hosts: AllowHosts, 
-    pub cors_allow_origin: Option<Vec<AccessControlAllowOrigin>>,
-    pub cors_max_age: Option<u32>,
-    pub cors_allow_headers: AccessControlAllowHeaders,
-    pub continue_on_invalid_cors: bool,
+    allow_hosts: AllowHosts,    
+    cors_allow_origin: Option<Vec<AccessControlAllowOrigin>>,
+    cors_max_age: Option<u32>,
+    cors_allow_headers: AccessControlAllowHeaders,
+    continue_on_invalid_cors: bool,
+}
+
+
+impl AccessControl {
+    /// Validate incoming request by http HOST
+    pub fn deny_host(&self, request: &hyper::Request<hyper::Body>) -> bool {
+        !hosts::is_host_valid(utils::read_header(request, "host"), &self.allow_hosts)
+    }
+
+    /// Validate incoming request by CORS origin 
+    pub fn deny_cors_origin(&self, request: &hyper::Request<hyper::Body>) -> bool {
+        let header = cors::get_cors_allow_origin(
+            utils::read_header(request, "origin"),
+            utils::read_header(request, "host"),
+            &self.cors_allow_origin,
+        )
+        .map(|origin| {
+            use self::cors::AccessControlAllowOrigin::*;
+            match origin {
+                Value(ref val) => {
+                    header::HeaderValue::from_str(val).unwrap_or_else(|_| header::HeaderValue::from_static("null"))
+                }
+                Null => header::HeaderValue::from_static("null"),
+                Any => header::HeaderValue::from_static("*"),
+            }
+        });
+        header == cors::AllowCors::Invalid && !self.continue_on_invalid_cors
+    }
+
+    /// Validate incoming request by CORS header
+    pub fn deny_cors_header(&self, request: &hyper::Request<hyper::Body>) -> bool {
+        let headers = request.headers().keys().map(|name| name.as_str());
+        let requested_headers = request
+            .headers()
+            .get_all("access-control-request-headers")
+            .iter()
+            .filter_map(|val| val.to_str().ok())
+            .flat_map(|val| val.split(", "))
+            .flat_map(|val| val.split(','));
+
+        let header = cors::get_cors_allow_headers(headers, requested_headers, &self.cors_allow_headers, |name| {
+            header::HeaderValue::from_str(name).unwrap_or_else(|_| header::HeaderValue::from_static("unknown"))
+        });
+        header == cors::AllowCors::Invalid && !self.continue_on_invalid_cors
+    }
 }
 
 impl Default for AccessControl {
@@ -22,6 +73,7 @@ impl Default for AccessControl {
     }
 }
 
+/// Convenience builder pattern 
 pub struct AccessControlBuilder {
     allow_hosts: AllowHosts, 
     cors_allow_origin: Option<Vec<AccessControlAllowOrigin>>,
